+++ PSScriptTools.psm1
@@ -1,10 +1,17 @@
+#enable verbose messaging in the psm1 file
+if ($myinvocation.line -match "-verbose") {
+    $VerbosePreference = "continue"
+}
+Write-Verbose "Loading public functions"
 
 Get-ChildItem -path $PSScriptRoot\functions\*.ps1 | foreach-object -process {
+    write-verbose $_.fullname
     . $_.FullName
 }
 
 #add ToDo options to the ISE or VS Code
 if ($psEditor) {
+    write-verbose "Defining VSCode additions"
     $sb = {
         Param(
             [Microsoft.PowerShell.EditorServices.Extensions.EditorContext]$context
@@ -20,6 +27,7 @@ if ($psEditor) {
 
 }
 elseif ($psise) {
+        write-verbose "Defining ISE additions"
     $action = {
 
         $prompt = "What do you need to do?"
@@ -28,9 +36,9 @@ elseif ($psise) {
         $todo = "# [$(Get-Date)] TODO: $item"
         $psise.CurrentFile.Editor.InsertText($todo)
         #jump cursor to the end
-        $psise.CurrentFile.editor.SetCaretPosition($psise.CurrentFile.Editor.CaretLine,$psise.CurrentFile.Editor.CaretColumn)
+        $psise.CurrentFile.editor.SetCaretPosition($psise.CurrentFile.Editor.CaretLine, $psise.CurrentFile.Editor.CaretColumn)
     }
 
     #add the action to the Add-Ons menu
-    $psISE.CurrentPowerShellTab.AddOnsMenu.Submenus.Add("ToDo",$Action,"Ctrl+Alt+2" ) | Out-Null
+    $psISE.CurrentPowerShellTab.AddOnsMenu.Submenus.Add("ToDo", $Action, "Ctrl+Alt+2" ) | Out-Null
 }